Refactor shutdown script to accept hours and seconds

+++ dynShutdown.ps1
@@ -1,13 +1,19 @@
 # ask user shutdown time
 # no value == 10 seconds to shutdown
-# value == value seconds to shutdoen
+# value == value seconds to shutdown
 # TODO: string("1:12:45") == (1*60^(2) + 12*60^(1) + 45*60^(0)) seconds to shutdown
 
 [CmdletBinding()]
 param (
     [Parameter()]
     [int]
-    $minutes = 0
+    $seconds = 0,
+    [int]
+    $minutes = 0,
+    [int]
+    $hours = 0,
+    [switch]
+    $keepConsole
 )
 
 function shtd {
@@ -15,19 +21,17 @@ function shtd {
     param (
         [Parameter()]
         [int]
-        $t
+        $t = 15
     )
 
-    shutdown.exe /s /t (60*$t);
-    $msg1="`n`n`nHerunterfahren in $($t) Minuten.`n`n`n"
+    shutdown.exe /s /t ($t);
+    $msg1="`n`n`nHerunterfahren in $($t/60) Minuten.`n`n`n";
     
     Write-Output $msg1;
-    
-    #[System.Windows.Forms.MessageBox]::Show("$msg1","Hinweis - Du Depp")
 }
 
 
-if ($minutes -eq 0) {
+if (($seconds -eq 0) -and ($minutes -eq 0) -and ($hours -eq 0)) {
     $time = Read-Host "In wie viel Minuten soll heruntergefahren werden?";
     if ($time -eq "")
     {
@@ -35,16 +39,15 @@ if ($minutes -eq 0) {
     }
     else
     {
-        shtd -t $time 
+        shtd -t $time;
     }
 }
 else
 {
-    shtd -t $minutes
+    shtd -t ($seconds + $minutes * 60 + $hours * 3600);
 }
 
 
 
-Start-Sleep -Seconds 3;
-
-Stop-Process -Name "pwsh" -Force;
+Start-Sleep -Seconds 2;
+if (!$keepConsole){Stop-Process -Name "pwsh" -Force;}
